+++ src/libtiledb.cpp
@@ -2001,7 +2001,10 @@ XPtr<tiledb::Array> libtiledb_array_open_at(XPtr<tiledb::Context> ctx, std::stri
     auto query_type = _string_to_tiledb_query_type(type);
     // get timestamp as seconds since epoch (plus fractional seconds, returns double), scale to millisec
     uint64_t ts_ms = static_cast<uint64_t>(std::round(tstamp.getFractionalTimestamp() * 1000));
-#if TILEDB_VERSION >= TileDB_Version(2,3,0)
+#if TILEDB_VERSION >= TileDB_Version(2,15,0)
+    auto ptr = new tiledb::Array(*ctx.get(), uri, query_type, tiledb::TemporalPolicy(tiledb::TimeTravel, ts_ms));
+    ptr->set_open_timestamp_end(ts_ms);
+#elif TILEDB_VERSION >= TileDB_Version(2,3,0)
     auto ptr = new tiledb::Array(*ctx.get(), uri, query_type);
     ptr->set_open_timestamp_end(ts_ms);
 #else